fix: lint

+++ tests/operations.rs
@@ -2347,9 +2347,8 @@ fn log0_with_gas_cost() {
     ];
     let topic_count = 0;
     let static_gas = gas_cost::LOG + gas_cost::PUSHN * (2 + topic_count);
-    let dynamic_gas = gas_cost::LOG * topic_count
-        + 8 * size as i64
-        + gas_cost::memory_expansion_cost(0, 32 as u32);
+    let dynamic_gas =
+        gas_cost::LOG * topic_count + 8 * size as i64 + gas_cost::memory_expansion_cost(0, 32_u32);
     let gas_needed = static_gas + dynamic_gas;
     run_program_assert_gas_exact(program, gas_needed as _);
 }
@@ -2369,9 +2368,8 @@ fn log1_with_gas_cost() {
     ];
     let topic_count = 1;
     let static_gas = gas_cost::LOG + gas_cost::PUSHN * (2 + topic_count);
-    let dynamic_gas = gas_cost::LOG * topic_count
-        + 8 * size as i64
-        + gas_cost::memory_expansion_cost(0, 32 as u32);
+    let dynamic_gas =
+        gas_cost::LOG * topic_count + 8 * size as i64 + gas_cost::memory_expansion_cost(0, 32_u32);
     let gas_needed = static_gas + dynamic_gas;
     run_program_assert_gas_exact(program, gas_needed as _);
 }
@@ -2392,9 +2390,8 @@ fn log2_with_gas_cost() {
     ];
     let topic_count = 2;
     let static_gas = gas_cost::LOG + gas_cost::PUSHN * (2 + topic_count);
-    let dynamic_gas = gas_cost::LOG * topic_count
-        + 8 * size as i64
-        + gas_cost::memory_expansion_cost(0, 32 as u32);
+    let dynamic_gas =
+        gas_cost::LOG * topic_count + 8 * size as i64 + gas_cost::memory_expansion_cost(0, 32_u32);
     let gas_needed = static_gas + dynamic_gas;
     run_program_assert_gas_exact(program, gas_needed as _);
 }
@@ -2416,9 +2413,8 @@ fn log3_with_gas_cost() {
     ];
     let topic_count = 3;
     let static_gas = gas_cost::LOG + gas_cost::PUSHN * (2 + topic_count);
-    let dynamic_gas = gas_cost::LOG * topic_count
-        + 8 * size as i64
-        + gas_cost::memory_expansion_cost(0, 32 as u32);
+    let dynamic_gas =
+        gas_cost::LOG * topic_count + 8 * size as i64 + gas_cost::memory_expansion_cost(0, 32_u32);
     let gas_needed = static_gas + dynamic_gas;
     run_program_assert_gas_exact(program, gas_needed as _);
 }
@@ -2441,9 +2437,8 @@ fn log4_with_gas_cost() {
     ];
     let topic_count = 4;
     let static_gas = gas_cost::LOG + gas_cost::PUSHN * (2 + topic_count);
-    let dynamic_gas = gas_cost::LOG * topic_count
-        + 8 * size as i64
-        + gas_cost::memory_expansion_cost(0, 32 as u32);
+    let dynamic_gas =
+        gas_cost::LOG * topic_count + 8 * size as i64 + gas_cost::memory_expansion_cost(0, 32_u32);
     let gas_needed = static_gas + dynamic_gas;
     run_program_assert_gas_exact(program, gas_needed as _);
 }
